convert python class to model
